separate backbone, compute_loss, fix tests
