+++ keras_hub/src/models/basnet/basnet_test.py
@@ -2,6 +2,7 @@ import pytest
 from keras import ops
 
 from keras_hub.src.models.basnet.basnet import BASNetImageSegmenter
+from keras_hub.src.models.basnet.basnet_backbone import BASNetBackbone
 from keras_hub.src.models.basnet.basnet_preprocessor import BASNetPreprocessor
 from keras_hub.src.models.resnet.resnet_backbone import ResNetBackbone
 from keras_hub.src.tests.test_case import TestCase
@@ -13,7 +14,7 @@ class BASNetTest(TestCase):
         self.labels = ops.concatenate(
             (ops.zeros((2, 32, 64, 1)), ops.ones((2, 32, 64, 1))), axis=1
         )
-        self.backbone = ResNetBackbone(
+        self.image_encoder = ResNetBackbone(
             input_conv_filters=[64],
             input_conv_kernel_sizes=[7],
             stackwise_num_filters=[64, 128, 256, 512],
@@ -21,11 +22,14 @@ class BASNetTest(TestCase):
             stackwise_num_strides=[1, 2, 2, 2],
             block_type="basic_block",
         )
+        self.backbone = BASNetBackbone(
+            image_encoder=self.image_encoder,
+            num_classes=1,
+        )
         self.preprocessor = BASNetPreprocessor()
         self.init_kwargs = {
             "backbone": self.backbone,
             "preprocessor": self.preprocessor,
-            "num_classes": 1,
         }
         self.train_data = (self.images, self.labels)
 
@@ -34,7 +38,7 @@ class BASNetTest(TestCase):
             cls=BASNetImageSegmenter,
             init_kwargs=self.init_kwargs,
             train_data=self.train_data,
-            expected_output_shape=[(2, 64, 64, 1)] * 8,
+            expected_output_shape=(2, 64, 64, 1),
         )
 
     @pytest.mark.large
@@ -47,10 +51,8 @@ class BASNetTest(TestCase):
 
     def test_end_to_end_model_predict(self):
         model = BASNetImageSegmenter(**self.init_kwargs)
-        outputs = model.predict(self.images)
-        self.assertAllEqual(
-            [output.shape for output in outputs], [(2, 64, 64, 1)] * 8
-        )
+        output = model.predict(self.images)
+        self.assertAllEqual(output.shape, (2, 64, 64, 1))
 
     @pytest.mark.skip(reason="disabled until preset's been uploaded to Kaggle")
     @pytest.mark.extra_large
@@ -60,5 +62,5 @@ class BASNetTest(TestCase):
                 cls=BASNetImageSegmenter,
                 preset=preset,
                 input_data=self.images,
-                expected_output_shape=[(2, 64, 64, 1)] * 8,
+                expected_output_shape=(2, 64, 64, 1),
             )
